Update jquery-v2.1.0.js
created some variables that will be defined later, as we write more code. then we want to define the DOM according to the sandbox, basically means where the server is local or host. define a local copy of jQuery but we need to initialize it but using our init function, if not we will get an error, set the cameCase function.