Update jquery-v2.1.0.js
first what we do is get any array-like objects and convert them into an array, which out this we would not be able to return "" into jQuery. then we run the get to return just one element(<div>DIV</div>) or get all the elements form the DOM, and then return all the elements in a clean array using slice.call( this ).  then push the newly created elements to the stack, merge any previous elements.