Update jquery-v2.1.0.js
we started breaking down the extend() function, as basically allow it to overwrite the first parenthesis that It is called to,  merging the contents of two or more objects together into the first object straight out of docs from jQuery. then we wrote what is variables we will be defining, this basically allows you to do the merge for extent, target is the array or the object.